do not use rigid bodies in derivatives calculations. use double and not floats in derivatives calculations
git-svn-id: https://svn.salilab.org/imp/trunk@9495 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/em/include/CoarseCCatIntervals.h
@@ -54,8 +54,8 @@ the same value before computing the values again
   \return the value of the cross correlation term: scalefac * (1-CCC)
    */
   float evaluate(DensityMap *em_map, SampledDensityMap *model_map,
-                 std::vector<float> &dvx, std::vector<float>&dvy,
-                 std::vector<float>&dvz, float scalefac, bool lderiv,
+                 std::vector<double> &dvx, std::vector<double>&dvy,
+                 std::vector<double>&dvz, float scalefac, bool lderiv,
                  unsigned long eval_interval);
 
   IMP_OBJECT_INLINE(CoarseCCatIntervals, {out << "CoarseCCatIntervals";}, {});
